Updated the support page
Just fixed a line at the beginning. -	The Live Chat link looks more like a static picture than a link you need to click to access the live chat. I would change that just so the “Live Chat” text is the hyperlink. Also this way we don’t have the same chat avatar on the same page